introduces cell ui
