+++ src/Facades/Cell.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace LaravelViews\Facades;
+
+use Illuminate\Support\Facades\Facade;
+
+class Cell extends Facade
+{
+    protected static function getFacadeAccessor()
+    {
+        return 'cell';
+    }
+}
