+++ src/LaravelViewsServiceProvider.php
@@ -12,6 +12,7 @@ use LaravelViews\Data\TableViewSearchData;
 use LaravelViews\UI\UI;
 use LaravelViews\UI\Variants;
 use LaravelViews\UI\Header;
+use LaravelViews\UI\Cell;
 use Illuminate\Support\ServiceProvider;
 use Illuminate\Support\Facades\Blade;
 use LaravelViews\Console\GridViewMakeCommand;
@@ -57,6 +58,9 @@ class LaravelViewsServiceProvider extends ServiceProvider
         $this->app->bind('ui', UI::class);
         $this->app->bind('header', function () {
             return new Header();
+        });
+        $this->app->bind('cell', function () {
+            return new Cell();
         });
     }
