+++ src/UI/Header.php
@@ -2,7 +2,7 @@
 
 namespace LaravelViews\UI;
 
-class Header
+class Header extends Cell
 {
     /** @var string Header's title to be shown */
     public $title;
@@ -13,16 +13,13 @@ class Header
     /** @var string Width the width of the table column */
     public $width;
 
-    /** @var string Breakpoint (min-width) from where column is displayed */
-    public $visibleBreakpoint;
-
     /**
      * Sets the header's title
      * @param string $title Header's title to be shown
      */
     public function title(string $title)
     {
-        $header = new static;
+        $header = $this->content($title);
         $header->title = $title;
 
         return $header;
@@ -58,33 +55,5 @@ class Header
         $this->width = $width;
 
         return $this;
-    }
-
-    /**
-     * Sets breakpoint from where column will be displayed
-     * @return Header
-     */
-    public function hideOnMobile(string $visibleBreakpoint = 'md')
-    {
-        $this->visibleBreakpoint = $visibleBreakpoint;
-
-        return $this;
-    }
-
-    /**
-     * Gets responsive table cell class names
-     * @return string
-     */
-    public function getResponsiveClassNames()
-    {
-        if (empty($this->visibleBreakpoint)) return '';
-
-        return match ($this->visibleBreakpoint) {
-            'sm' => 'hidden sm:table-cell',
-            'lg' => 'hidden lg:table-cell',
-            'xl' => 'hidden xl:table-cell',
-            '2xl' => 'hidden 2xl:table-cell',
-            default => 'hidden md:table-cell'
-        };
     }
 }
